requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.2
accelerate>=0.26.0
bitsandbytes>=0.41.0
timm>=1.0.0

# Transformers requirements
sentencepiece
protobuf

# Data handling and visualization
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities
pyyaml>=6.0
scikit-learn>=1.2.0
tqdm>=4.65.0
tensorboard>=2.13.0
einops>=0.6.0

# Stage 1 & 2 additions
safetensors>=0.4.0

# Phase 3 & 4 additions
nltk>=3.8.1
rouge>=1.0.1
flask>=2.3.0
tensorboard-data-server>=0.7.0

# Development tools
# pytest>=7.3.1
# black>=23.3.0
# isort>=5.12.0
# flake8>=6.0.0
# mypy>=1.3.0

# Optional dependencies - comment out if installation issues
# flash-attn>=2.5.0  # Requires CUDA setup
# xformers>=0.0.23  # Efficient attention mechanisms
# deepspeed>=0.12.0  # Multi-GPU training